Request util upgrade
Previously, interceptors (onRequest, onRequestError, onResponse, onResponseError) and HTTP headers in options could not be used when initializing a request.

This created an inconvenience, because of which it was impossible to simply repeat the request if the previous response from the server returned an error.

In this update, at the time the request is actually sent, HTTP headers and interceptors from the options and from the request context will merge together.